pyproject.toml
[project]
name = "ai-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.27",
    "faker>=40.1.0",
    "fastapi==0.115.0",
    "joblib==1.4.2",
    "kafka-python-ng>=2.2.3",
    "matplotlib==3.9.2",
    "mlflow>=3.8.1",
    "pandas==2.2.2",
    "psycopg2-binary>=2.9.11",
    "pydantic==2.8.2",
    "pydantic-settings>=2.12.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "scikit-learn==1.5.0",
    "streamlit==1.37.1",
    "uvicorn[standard]==0.30.6",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["api", "consumer", "producer", "dashboard", "ml", "shared"]

[tool.pytest.ini_options]
testpaths = ["tests"]

